+++ drizzle/0005_tan_wild_child.sql
@@ -1,0 +1,15 @@
+CREATE TYPE "public"."grading_status" AS ENUM('complete', 'pending', 'failed');--> statement-breakpoint
+CREATE TABLE "otps" (
+	"id" uuid PRIMARY KEY DEFAULT gen_random_uuid() NOT NULL,
+	"key" text NOT NULL,
+	"code" text NOT NULL,
+	"expires_at" timestamp NOT NULL,
+	"created_at" timestamp DEFAULT now() NOT NULL,
+	CONSTRAINT "otps_key_unique" UNIQUE("key")
+);
+--> statement-breakpoint
+ALTER TABLE "quiz_results" ADD COLUMN "grading_status" "grading_status" DEFAULT 'complete' NOT NULL;--> statement-breakpoint
+ALTER TABLE "user_answers" ADD COLUMN "is_correct" boolean;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "is_verified" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "users" DROP COLUMN "password_reset_token_hash";--> statement-breakpoint
+ALTER TABLE "users" DROP COLUMN "password_reset_token_expires_at";
