+++ src/controllers/authEmailController.ts
@@ -2,7 +2,7 @@ import { NextFunction, Request, Response } from 'express'
 
 import { successResponse } from '../helpers/apiResponse'
 import { AuthRequest } from '../middlewares/authMiddleware'
-import authEmailService from '../services/authEmailService'
+import authEmailService from '../services/auth-email.service'
 
 const COOKIE_ACCESS_TOKEN_OPTIONS = {
   httpOnly: true,
@@ -35,7 +35,7 @@ const confirmRegistration = async (req: Request, res: Response, next: NextFuncti
     res.cookie('accessToken', accessToken, COOKIE_ACCESS_TOKEN_OPTIONS)
     res.cookie('refreshToken', refreshToken, COOKIE_REFRESH_TOKEN_OPTIONS)
 
-    return res.redirect(`${process.env.FRONTEND_URL ?? 'http://localhost:5173'}/app/dashboard`)
+    return res.redirect(`${process.env.FRONTEND_URL ?? 'http://localhost:5173'}/app/analytics`)
   } catch (error) {
     next(error)
   }
@@ -49,7 +49,7 @@ const login = async (req: Request, res: Response, next: NextFunction) => {
     res.cookie('accessToken', accessToken, COOKIE_ACCESS_TOKEN_OPTIONS)
     res.cookie('refreshToken', refreshToken, COOKIE_REFRESH_TOKEN_OPTIONS)
 
-    return res.redirect(`${process.env.FRONTEND_URL ?? 'http://localhost:5173'}/app/dashboard`)
+    return res.redirect(`${process.env.FRONTEND_URL ?? 'http://localhost:5173'}/app/analytics`)
   } catch (error) {
     next(error)
   }
