+++ src/controllers/quizController.ts
@@ -4,9 +4,10 @@ import { successResponse } from '../helpers/apiResponse'
 import { AppError } from '../helpers/AppError'
 import { getAuthUserId } from '../helpers/utils/authUtils'
 import { parseS3Url } from '../helpers/utils/quizUtils'
-import { invokeQuizGenerator } from '../services/invokeQuizGenerator'
-import notionQuizService from '../services/notionQuizService'
-import profileService from '../services/profileService'
+import { invokeQuizGenerator } from '../services/helpers/invokeQuizGenerator'
+import notionQuizService from '../services/notion-quiz.service'
+import profileService from '../services/profile.service'
+import { getQuizResult, submitQuiz } from '../services/quiz-submission.service'
 import { getPublicQuizByToken } from '../services/quiz.service'
 import { setQuizSharing } from '../services/quiz.service'
 import {
@@ -14,7 +15,6 @@ import {
   getJobById,
   getQuizById,
   getQuizzes,
-  submitQuiz,
   updateQuizById,
 } from '../services/quiz.service'
 import type {
@@ -258,6 +258,28 @@ export const submitQuizController = async (req: Request, res: Response, next: Ne
     next(error)
   }
 }
+
+export const getQuizResultController = async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    const userId = getAuthUserId(req)
+
+    const rawId = req.params.id
+    const quizId = typeof rawId === 'string' ? rawId : rawId?.[0]
+    if (!quizId) {
+      throw new AppError('Invalid quiz id', 400, 'VALIDATION_ERROR')
+    }
+
+    const data = await getQuizResult(quizId, userId)
+    if (!data) {
+      throw new AppError('Result not found', 404, 'NOT_FOUND')
+    }
+
+    res.status(200).json(successResponse('Quiz result retrieved', data))
+  } catch (error) {
+    next(error)
+  }
+}
+
 export const enableSharingController = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const userId = getAuthUserId(req)
