+++ src/controllers/userProfileControllers.ts
@@ -4,7 +4,7 @@ import { successResponse } from '../helpers/apiResponse'
 import { AuthRequest } from '../middlewares/authMiddleware'
 import User from '../models/user.model'
 import userProfile from '../models/userProfile.model'
-import UserProfileImageService from '../services/userProfileImageService'
+import UserProfileImageService from '../services/user-profile-image.service'
 
 export const getUserProfile = async (req: AuthRequest, res: Response, next: NextFunction) => {
   try {
