+++ src/prompts/index.ts
@@ -1,0 +1,3 @@
+export { buildQuizSystemPrompt } from './quiz-generation.prompt'
+export { OPEN_ENDED_GRADING_SYSTEM_PROMPT } from './open-ended-grading.prompt'
+export { FEEDBACK_SYSTEM_PROMPT } from './feedback.prompt'
