+++ src/routes/quiz.routes.ts
@@ -8,6 +8,7 @@ import {
   getJobStatusController,
   getPublicQuizController,
   getQuizByIdController,
+  getQuizResultController,
   getQuizzesController,
   patchQuizByIdController,
   submitQuizController,
@@ -295,6 +296,32 @@ router.patch('/quizzes/:id/share/disable', authMiddleware, disableSharingControl
  *         description: Quiz not found
  */
 router.post('/quizzes/:id/submit', authMiddleware, validate(SubmitQuizSchema), submitQuizController)
+
+/**
+ * @openapi
+ * /quizzes/{id}/result:
+ *   get:
+ *     tags:
+ *       - Quiz
+ *     security:
+ *       - cookieAuth: []
+ *     summary: Get a quiz result with open-ended grading status and per-answer verdicts
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: string
+ *           format: uuid
+ *     responses:
+ *       200:
+ *         description: Result with grading status and verdicts
+ *       401:
+ *         description: Not authenticated
+ *       404:
+ *         description: Result not found
+ */
+router.get('/quizzes/:id/result', authMiddleware, getQuizResultController)
 
 /**
  * @openapi
