+++ src/services/auth-email.service.ts
@@ -3,9 +3,9 @@ import crypto from 'crypto'
 import bcrypt from 'bcryptjs'
 import { eq } from 'drizzle-orm'
 
-import authService from './authService'
-import emailService from './emailService'
-import userService from './userService'
+import authService from './auth.service'
+import emailService from './clients/email.client'
+import userService from './user.service'
 import { db } from '../database/database'
 import { users } from '../database/schema'
 import { AppError } from '../helpers/AppError'
