+++ src/services/auth.service.ts
@@ -1,10 +1,10 @@
 import { eq } from 'drizzle-orm'
 import jwt from 'jsonwebtoken'
 
-import integrationService from './integrationService'
-import notionService from './notionService'
-import profileService from './profileService'
-import userService from './userService'
+import integrationService from './integration.service'
+import notionService from './notion.service'
+import profileService from './profile.service'
+import userService from './user.service'
 import { db } from '../database/database'
 import { users } from '../database/schema'
 import { AppError } from '../helpers/AppError'
