+++ src/services/notion-quiz.service.ts
@@ -2,9 +2,9 @@ import { randomUUID } from 'crypto'
 
 import { PutObjectCommand } from '@aws-sdk/client-s3'
 
-import { invokeQuizGenerator } from './invokeQuizGenerator'
-import notionService from './notionService'
-import { s3Client, s3BucketName } from './s3Client'
+import { s3Client, s3BucketName } from './clients/s3.client'
+import { invokeQuizGenerator } from './helpers/invokeQuizGenerator'
+import notionService from './notion.service'
 import { AppError } from '../helpers/AppError'
 import type { QuestionType } from '../types/questionTypes'
